Better multi-platform example
